update u32 to usize to avoid overflow when defining output vector/matrix sizes

+++ src/linalg/add.rs
@@ -92,9 +92,9 @@ impl Matrix {
         let f: CudaFunction = self.get_cached_kernel("cuScalarMatAdd", SCALARMATADD)?;
         let stream: Arc<CudaStream> = self.data.context().default_stream();
         let mut builder: LaunchArgs = stream.launch_builder(&f);
-        let n_rows: u32 = self.n_rows as u32;
-        let n_cols: u32 = self.n_cols as u32;
-        let out: Vec<f32> = vec![0.0; (n_rows * n_cols) as usize];
+        let n_rows: usize = self.n_rows;
+        let n_cols: usize = self.n_cols;
+        let out: Vec<f32> = vec![0.0; n_rows * n_cols];
         let mut out_dev: CudaSlice<f32> = stream.clone_htod(&out)?;
         builder.arg(&s);
         builder.arg(&self.data);
@@ -104,8 +104,8 @@ impl Matrix {
         let cfg = LaunchConfig {
             block_dim: (BLOCK_SIZE, BLOCK_SIZE, 1),
             grid_dim: (
-                (n_cols + BLOCK_SIZE - 1) / BLOCK_SIZE,
-                (n_rows + BLOCK_SIZE - 1) / BLOCK_SIZE,
+                ((n_cols as u32) + BLOCK_SIZE - 1) / BLOCK_SIZE,
+                ((n_rows as u32) + BLOCK_SIZE - 1) / BLOCK_SIZE,
                 1,
             ),
             shared_mem_bytes: 0,
@@ -113,7 +113,7 @@ impl Matrix {
         unsafe {
             let _ = builder.launch(cfg);
         };
-        Ok(Self::new(out_dev, n_rows as usize, n_cols as usize)?)
+        Ok(Self::new(out_dev, n_rows, n_cols)?)
     }
 
     pub fn elementwisematadd(self: &Self, b: &Self) -> Result<Self, Box<dyn Error>> {
@@ -126,9 +126,9 @@ impl Matrix {
         let f: CudaFunction = self.get_cached_kernel("cuElementwiseMatAdd", ELEMENTWISEMATADD)?;
         let stream: Arc<CudaStream> = self.data.context().default_stream();
         let mut builder: LaunchArgs = stream.launch_builder(&f);
-        let n_rows: u32 = self.n_rows as u32;
-        let n_cols: u32 = self.n_cols as u32;
-        let out: Vec<f32> = vec![0.0; (n_rows * n_cols) as usize];
+        let n_rows: usize = self.n_rows;
+        let n_cols: usize = self.n_cols;
+        let out: Vec<f32> = vec![0.0; n_rows * n_cols];
         let mut out_dev: CudaSlice<f32> = stream.clone_htod(&out)?;
         builder.arg(&self.data);
         builder.arg(&b.data);
@@ -138,8 +138,8 @@ impl Matrix {
         let cfg = LaunchConfig {
             block_dim: (BLOCK_SIZE, BLOCK_SIZE, 1),
             grid_dim: (
-                (n_cols + BLOCK_SIZE - 1) / BLOCK_SIZE,
-                (n_rows + BLOCK_SIZE - 1) / BLOCK_SIZE,
+                ((n_cols as u32) + BLOCK_SIZE - 1) / BLOCK_SIZE,
+                ((n_rows as u32) + BLOCK_SIZE - 1) / BLOCK_SIZE,
                 1,
             ),
             shared_mem_bytes: 0,
@@ -147,7 +147,7 @@ impl Matrix {
         unsafe {
             let _ = builder.launch(cfg);
         };
-        Ok(Self::new(out_dev, n_rows as usize, n_cols as usize)?)
+        Ok(Self::new(out_dev, n_rows, n_cols)?)
     }
 
     pub fn rowmatadd(self: &Self, b: &Self) -> Result<Self, Box<dyn Error>> {
@@ -160,9 +160,9 @@ impl Matrix {
         let f: CudaFunction = self.get_cached_kernel("cuRowMatAdd", ROWMATADD)?;
         let stream: Arc<CudaStream> = self.data.context().default_stream();
         let mut builder: LaunchArgs = stream.launch_builder(&f);
-        let n_rows: u32 = self.n_rows as u32;
-        let n_cols: u32 = self.n_cols as u32;
-        let out: Vec<f32> = vec![0.0; (n_rows * n_cols) as usize];
+        let n_rows: usize = self.n_rows;
+        let n_cols: usize = self.n_cols;
+        let out: Vec<f32> = vec![0.0; n_rows * n_cols];
         let mut out_dev: CudaSlice<f32> = stream.clone_htod(&out)?;
         builder.arg(&self.data);
         builder.arg(&b.data);
@@ -172,8 +172,8 @@ impl Matrix {
         let cfg = LaunchConfig {
             block_dim: (BLOCK_SIZE, BLOCK_SIZE, 1),
             grid_dim: (
-                (n_cols + BLOCK_SIZE - 1) / BLOCK_SIZE,
-                (n_rows + BLOCK_SIZE - 1) / BLOCK_SIZE,
+                ((n_cols as u32) + BLOCK_SIZE - 1) / BLOCK_SIZE,
+                ((n_rows as u32) + BLOCK_SIZE - 1) / BLOCK_SIZE,
                 1,
             ),
             shared_mem_bytes: 0,
@@ -181,7 +181,7 @@ impl Matrix {
         unsafe {
             let _ = builder.launch(cfg);
         };
-        Ok(Self::new(out_dev, n_rows as usize, n_cols as usize)?)
+        Ok(Self::new(out_dev, n_rows, n_cols)?)
     }
 
     pub fn colmatadd(self: &Self, b: &Self) -> Result<Self, Box<dyn Error>> {
@@ -194,9 +194,9 @@ impl Matrix {
         let f: CudaFunction = self.get_cached_kernel("cuColMatAdd", COLMATADD)?;
         let stream: Arc<CudaStream> = self.data.context().default_stream();
         let mut builder: LaunchArgs = stream.launch_builder(&f);
-        let n_rows: u32 = self.n_rows as u32;
-        let n_cols: u32 = self.n_cols as u32;
-        let out: Vec<f32> = vec![0.0; (n_rows * n_cols) as usize];
+        let n_rows: usize = self.n_rows;
+        let n_cols: usize = self.n_cols;
+        let out: Vec<f32> = vec![0.0; n_rows * n_cols];
         let mut out_dev: CudaSlice<f32> = stream.clone_htod(&out)?;
         builder.arg(&self.data);
         builder.arg(&b.data);
@@ -206,8 +206,8 @@ impl Matrix {
         let cfg = LaunchConfig {
             block_dim: (BLOCK_SIZE, BLOCK_SIZE, 1),
             grid_dim: (
-                (n_cols + BLOCK_SIZE - 1) / BLOCK_SIZE,
-                (n_rows + BLOCK_SIZE - 1) / BLOCK_SIZE,
+                ((n_cols as u32) + BLOCK_SIZE - 1) / BLOCK_SIZE,
+                ((n_rows as u32) + BLOCK_SIZE - 1) / BLOCK_SIZE,
                 1,
             ),
             shared_mem_bytes: 0,
@@ -215,7 +215,7 @@ impl Matrix {
         unsafe {
             let _ = builder.launch(cfg);
         };
-        Ok(Self::new(out_dev, n_rows as usize, n_cols as usize)?)
+        Ok(Self::new(out_dev, n_rows, n_cols)?)
     }
 }
